🔨Refactor: 로그인 및 회원가입 controller -> AuthController로 통합
